+++ CHANGELOG.rst
@@ -4,11 +4,13 @@ Changelog
 0.16.0 (in development)
 -----------------------
 
+* Add new optional argument ``--comparison-table`` to :command:`train-cnv-caller` and :command:`test-cnv-caller` commands.
 * Improve CNV caller for CYP2A6, CYP2B6, CYP2D6, CYP2E1, GSTM1, SLC22A2, SULT1A1, UGT1A4, UGT2B15, UGT2B17.
 * Add new CNV call for CYP2A6: ``Unknown1`` and ``Hybrid7``.
 * Add new CNV calls for CYP2B6: ``Tandem``, ``PartialDup``, ``PartialDup2``, ``PseudogeneDel``.
 * Add new CNV call for CYP2D6: ``Del,Tandem3``.
 * Add new CNV calls for CYP2E1: ``Normal,Del`` and ``Dup,PartialDup``.
+* Add new CNV call for SLC22A2: ``Intron9DeletionHom``.
 * Add new CNV call for SULT1A1: ``Unknown2``.
 * Add new CNV call for UGT1A4: ``Intron1DeletionAHom``.